Encapsulate module for pool, refactoring
