+++ variables.tf
@@ -553,29 +553,29 @@ variable "runner_architecture" {
   }
 }
 
-variable "simple_pool_lambda_timeout" {
-  description = "Time out for the simple pool lambda lambda in seconds."
+variable "pool_lambda_timeout" {
+  description = "Time out for the pool lambda lambda in seconds."
   type        = number
   default     = 60
 }
 
-variable "simple_pool_runner_owner" {
-  description = "The simple pool wil deploy runners to an GitHub org, set this value to the org to which you want the runners deployed. Repo level is not supported."
+variable "pool_runner_owner" {
+  description = "The pool wil deploy runners to an GitHub org, set this value to the org to which you want the runners deployed. Repo level is not supported."
   type        = string
   default     = null
 }
 
-variable "simple_pool_reserved_concurrent_executions" {
+variable "pool_lambda_reserved_concurrent_executions" {
   description = "Amount of reserved concurrent executions for the scale-up lambda function. A value of 0 disables lambda from being triggered and -1 removes any concurrency limitations."
   type        = number
   default     = 1
 }
 
-variable "simple_pool_config" {
+variable "pool_config" {
   description = "The configuration for updating the pool. The `pool_size` to adjust to by the events triggered by the the `schedule_expression. For example you can configure a cron expression for week days to adjust the pool to 10 and another expression for the weekend to adjust the pool to 1."
   type = list(object({
     schedule_expression = string
-    pool_size           = number
+    size                = number
   }))
   default = []
 }
